# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "AgentClinic API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Local backend"
  } ],
  "paths" : {
    "/api/agents" : {
      "get" : {
        "tags" : [ "Agents" ],
        "operationId" : "listAgents",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AgentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Agents" ],
        "operationId" : "createAgent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAgentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Агент зарегистрирован",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибки валидации; errors: поле -> сообщение",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409" : {
            "description" : "Имя уже занято; errors.name",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/health" : {
      "get" : {
        "tags" : [ "Health" ],
        "operationId" : "health",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{id}" : {
      "get" : {
        "tags" : [ "Agents" ],
        "operationId" : "getAgent",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Агент не найден",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "CreateAgentRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 60,
            "minLength" : 2
          },
          "model" : {
            "type" : "string",
            "maxLength" : 60,
            "minLength" : 0
          },
          "vendor" : {
            "type" : [ "string", "null" ],
            "maxLength" : 60,
            "minLength" : 0
          },
          "description" : {
            "type" : [ "string", "null" ],
            "maxLength" : 1000,
            "minLength" : 0
          }
        },
        "required" : [ "model", "name" ]
      },
      "AgentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "vendor" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "description", "id", "model", "name", "vendor" ]
      },
      "ApiProblem" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "status", "title" ]
      },
      "HealthResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          }
        },
        "required" : [ "status" ]
      }
    }
  }
}